refactor: separate slices into different files and put them in a folder
