+++ react-ts-app/src/components/Store/Slices/createdCardsSlice.ts
@@ -1,0 +1,20 @@
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { INewCard } from "../../../types/types";
+
+interface ICreatedCardsState {
+  cards: INewCard[];
+}
+
+export const createdCardsSlice = createSlice({
+  name: "createdCards",
+  initialState: {
+    cards: [],
+  } as ICreatedCardsState,
+  reducers: {
+    addNewCard: (state, action: PayloadAction<INewCard>) => {
+      state.cards.push(action.payload);
+    }
+  }
+})
+
+export const { addNewCard } = createdCardsSlice.actions;
